+++ com.unity.render-pipelines.high-definition/Runtime/Water/WaterSystemDef.cs
@@ -101,9 +101,11 @@ namespace UnityEngine.Rendering.HighDefinition
 
         public Vector4 _FoamJacobianLambda;
 
-        public Vector2 _PaddinwW0;
-        public int _PaddinwW1;
+        // Offsets used to guarantee the coherence between the different simulation resolutions
+        public int _WaterRefSimRes;
+        public float _WaterSpectrumOffset;
         public int _WaterSampleOffset;
+        public int _WaterBandCount;
     }
 
     [GenerateHLSL(needAccessors = false, generateCBuffer = true)]